[fix] login error

+++ frontend/petclinicmock_frontend/src/main.js
@@ -32,6 +32,8 @@ axios.interceptors.response.use(
       if (error.response) {
           switch (error.response.status) {
               case 401:
+                  if(router.currentRoute.path === '/login')
+                    break;
                   // 返回 401 清除token信息并跳转到登录页面
                   window.alert("尚未登录或已过期");
                   localStorage.clear();
